Moved SignedIn and AuthUi to UserAuth package
