+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SignedInActivity"></activity>
+        <activity android:name=".UserAuth.SignedInActivity"></activity>
     </application>
 
 </manifest>
