+++ app/src/main/res/layout/activity_signed_in.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.studios.juke.juke.SignedInActivity">
+    tools:context="com.studios.juke.juke.UserAuth.SignedInActivity">
 
 </android.support.constraint.ConstraintLayout>
